Trunk: newlib now compiles and links with all libraries, examples and the unitTest.
git-svn-id: svn://localhost/public/MoSync/trunk@813 6ad7a17e-d552-4a1e-ace4-f958fe14b117

+++ libs/newlib/libc/sys/mosync/maheap.c
@@ -115,7 +115,7 @@ void ansi_heap_init_crt0(char *start, int length)
 //				malloc
 //****************************************
 
-void * malloc(int size)
+void * malloc(size_t size)
 {
 	void *result;
 
@@ -151,9 +151,9 @@ void * malloc(int size)
 //				calloc
 //****************************************
 
-void * calloc(int num_elem, int size_elem)
+void * calloc(size_t num_elem, size_t size_elem)
 {
-	int size = num_elem * size_elem;
+	size_t size = num_elem * size_elem;
 	void *mem;
 
 	MASTD_HEAP_LOG("calloc(%i, %i)\n", num_elem, size_elem);
@@ -195,7 +195,7 @@ void free(void *mem)
 //****************************************
 //				fugly realloc
 //****************************************
-void* realloc(void* old, int size) {
+void* realloc(void* old, size_t size) {
 	MASTD_HEAP_LOG("realloc(0x%08X, %i)\n", (int)old, size);
 
 	return gReallocHook(old, size);
